Implement alloc_create
This function will create a new allocator with a maximum memory
limit, which will be aligned up to 8 bytes.

It will add a first free block, and a last block. The last block
is always inuse, preventing some block being merged outside the
valid memory. Since the size is aligned to 8 bytes, its least
significant bit is used to represent if it is inuse.