Fix wrong implementation of binary search tree
The left son of a node in binary search tree is not necessary the
biggest element in the left subtree, so it is not correct to check
if the left son equal to or bigger than the target to decide if we
should go left. Instead, we can use a variable to store the smallest
element that conforms to the requirement on the path we visited and
use the current node to decide whether we should go left or right,
until the size of current node is exactly the same as the target or
it is NULL. Furthermore, it is easy to notice that the smallest
elements is the last element that conforms to the requirement.